@@ -1,30 +1,26 @@
 buildscript {
     repositories {
-        maven { url 'https://maven.fabric.io/public' }
-    }
-
-    dependencies {
-        classpath 'io.fabric.tools:gradle:1.27.1'
+        google()
     }
 }
 
 apply plugin: 'com.android.application'
-apply plugin: 'io.fabric'
+apply plugin: 'com.google.firebase.crashlytics'
+apply plugin: 'com.google.gms.google-services'
 
 repositories {
     maven { url 'https://maven.google.com' }
-    maven { url 'https://maven.fabric.io/public' }
     maven { url 'https://jitpack.io' }
 }
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '26.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.0'
 
     defaultConfig {
         applicationId project.property('demo_applicationId')
-        minSdkVersion 15
-        targetSdkVersion 28
+        minSdkVersion 16
+        targetSdkVersion 29
         versionCode project.property('versionCode').toString().toInteger()
         versionName project.property('versionName')
         vectorDrawables {
@@ -61,23 +57,19 @@ android {
 }
 
 dependencies {
-    compile project(':webimclientsdkandroid')
-    compile 'com.astuetz:pagerslidingtabstrip:1.0.1'
-    compile 'com.github.bumptech.glide:glide:3.7.0'
-    compile 'com.github.bumptech.glide:okhttp3-integration:1.4.0@aar'
-    compile 'de.hdodenhof:circleimageview:1.3.0'
-    compile 'jp.wasabeef:glide-transformations:2.0.1'
-    compile 'com.android.support:appcompat-v7:28.0.0'
-    compile 'com.android.support:design:28.0.0'
-    compile 'com.google.firebase:firebase-messaging:17.3.4'
-    compile 'com.firebase:firebase-jobdispatcher:0.8.5'
-    compile('com.crashlytics.sdk.android:crashlytics:2.6.3@aar') {
-        transitive = true
-    }
-    compile 'com.github.thijsk:TouchImageView:v1.3.1'
-    compile 'com.android.support:recyclerview-v7:28.0.0'
-    compile 'com.android.support.constraint:constraint-layout:1.1.3'
-    compile 'com.android.support:cardview-v7:28.0.0'
+    implementation project(':webimclientsdkandroid')
+    implementation 'com.astuetz:pagerslidingtabstrip:1.0.1'
+    implementation 'com.github.bumptech.glide:glide:3.7.0'
+    implementation 'com.github.bumptech.glide:okhttp3-integration:1.4.0@aar'
+    implementation 'de.hdodenhof:circleimageview:3.1.0'
+    implementation 'jp.wasabeef:glide-transformations:2.0.1'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.1.0'
+    implementation 'com.google.firebase:firebase-messaging:20.2.0'
+    implementation 'com.firebase:firebase-jobdispatcher:0.8.5'
+    implementation 'com.google.firebase:firebase-crashlytics:17.0.1'
+    implementation 'com.github.thijsk:TouchImageView:v1.3.1'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.cardview:cardview:1.0.0'
 }
-
-apply plugin: 'com.google.gms.google-services'
