@@ -1,8 +1,8 @@
 package com.webimapp.android.demo.client.util;
 
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public abstract class EndlessScrollListener extends RecyclerView.OnScrollListener {
