@@ -4,8 +4,8 @@ package com.webimapp.android.sdk;
  * Created by Nikita Lazarev-Zubov on 13.12.17
  */
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.net.URL;
 import java.util.Map;
