@@ -1,7 +1,7 @@
 package com.webimapp.android.sdk;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 /**
  * Abstracts a chat operator
