@@ -1,11 +1,10 @@
 package com.webimapp.android.sdk.impl;
 
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.webimapp.android.sdk.FAQ;
 import com.webimapp.android.sdk.FAQCategory;
