@@ -1,6 +1,6 @@
 package com.webimapp.android.sdk.impl;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.List;
 import java.util.Set;
