@@ -1,6 +1,6 @@
 package com.webimapp.android.sdk.impl;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.webimapp.android.sdk.Message;
 import com.webimapp.android.sdk.MessageTracker;
