@@ -1,6 +1,6 @@
 package com.webimapp.android.sdk.impl;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 public interface MessageComposingHandler {
     void setComposingMessage(@Nullable String draftMessage);
