@@ -1,6 +1,6 @@
 package com.webimapp.android.sdk.impl.backend;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public interface SessionParamsListener {
     void onSessionParamsChanged(@NonNull String visitorJson,
