@@ -1,6 +1,6 @@
 package com.webimapp.android.sdk.impl.items;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.gson.annotations.SerializedName;
 
